+++ poll/src/App.svelte
@@ -8,17 +8,6 @@
 
 	let tabs: Array<string> = ['Create Polls', 'View Polls'];
 	
-	// data
-	let polls = [
-    {
-      id: 1,
-      question: 'Python or JavaScript?',
-      answerA: 'Python',
-      answerB: 'JavaScript',
-      votesA: 9,
-      votesB: 15,
-    },
-  ];
 
 
 	let currTab: string = "Create Polls";
@@ -29,24 +18,21 @@
 
 	const handleAdd = (e) => {
 		currTab = 'View Polls';
-		polls = [e.detail, ...polls];
 	}
 
-	const handleVote = (e) => {
-		// let id = e.detail.id;
-		// let option = e.detail.option;
-		const {option, id} = e.detail;
-		let copiedPolls = [...polls];
-		let upvotedPoll = copiedPolls.find(poll => poll.id == id);
-		if(option === 'a'){
-			upvotedPoll.votesA++;
-		}
-		if(option === 'b'){
-			upvotedPoll.votesB++;
-		}
-		polls = copiedPolls;
+	// const handleVote = (e) => {
+	// 	const {option, id} = e.detail;
+	// 	let copiedPolls = [...polls];
+	// 	let upvotedPoll = copiedPolls.find(poll => poll.id == id);
+	// 	if(option === 'a'){
+	// 		upvotedPoll.votesA++;
+	// 	}
+	// 	if(option === 'b'){
+	// 		upvotedPoll.votesB++;
+	// 	}
+	// 	polls = copiedPolls;
 		
-	}
+	// }
 	// let logger = (val) =>{
     //     console.log(val);
     // }
@@ -61,7 +47,7 @@
 	{#if currTab === "Create Polls"}
 		<PollsCreate on:add={handleAdd}/>
 	{:else}
-		<PollsView {polls} on:vote={handleVote}/>
+		<PollsView/>
 	{/if}
 </main>
